Reverted some breaking changes made in commit 077a6258a950843368b59959bfc762a98fc37b3d, and later. The change in this commit made `jupyterhub_traefik/install.py` not download any versions of traefik, consul or etcd where the checksums were not present.
Further, (and this was a change made in a later commit
faa2832437bf9a2bcd0023a24c34fea150d5d4c1), the default version of etcd was
updated to 3.4.15 and although this was present in the checksums, for some
reason it did not install properly in the github workflow.

Finally, updated some assert statements in tests/proxytest.py, to instead
run `assert_equal(val, cmp)`, which makes the pytest error message easier to
follow.