Enable TLS termination and use of specific entrypoints on the traefik routers.
As mentioned in
https://github.com/jupyterhub/traefik-proxy/pull/133#issuecomment-868431999
configurable options have been added to the base TraefikProxy class:-

 - TraefikProxy.traefik_tls
    Setting this to `True ` will set the traefik tls option on each
    jupyterhub-configured traefik router
 - TraefikProxy.default_entrypoint
    A string that should match the traefik entryPoint, for which each
    jupyterhub-configured service will be attached.
 - TraefikProxy.traefik_entrypoints
    A list of strings, where each entrypoint should match traefik-configured
    entrypoints. If empty, then just use the default_entrypoint

I'm not sure if this is the best way to do this to be honest. It looks like
jupyterhub's `bind_url` predicates use of a single port anyway. So, is only a
default_entrypoint necessary? Maybe these configuration options aren't required
and should just be figured out from a combination of the other jupyterhub
configuration variables (e.g. is bind_url http or https), and pulled from
existing jupyterhub configuration variables anyway?